update winit to version 0.24 to allow building on mac m1

+++ dotrix_core/src/input.rs
@@ -31,7 +31,7 @@ pub enum Button {
     /// Middle mouse button
     MouseMiddle,
     /// Mouse button by the code
-    MouseOther(u8),
+    MouseOther(u16),
 }
 
 /// State of a button
